release actions (#152)
* release actions

* bug fixeds

* actions fix

+++ apps/wallet/instance/app/src/main/java/com/tonapps/tonkeeper/core/entities/TransferEntity.kt
@@ -93,7 +93,7 @@ data class TransferEntity(
 
     private val coins: org.ton.block.Coins
         get() {
-            return org.ton.block.Coins.ofNano(amount.toLong())
+            return org.ton.block.Coins.ofNano(amount.toBigInteger())
         }
 
     private fun getCommentForwardPayload(
